Move MobileNavItem to own file, use MUI icons in MobileNav
