+++ src/components/nav/mobile-nav.js
@@ -1,62 +1,16 @@
 // External Dependencies
-import PropTypes from 'prop-types';
 import React from 'react';
-import { Link } from 'gatsby';
-
-// Icons
-import {
-  MdCardGiftcard as CardGiftCard,
-  MdFolderOpen as FolderOpen,
-  MdInfoOutline as InfoOutline,
-  MdPersonOutline as PersonOutline,
-} from 'react-icons/md';
+import CardGiftCard from '@material-ui/icons/CardGiftCard';
+import FolderOpen from '@material-ui/icons/FolderOpen';
+import InfoOutlined from '@material-ui/icons/InfoOutlined';
+import PersonOutlined from '@material-ui/icons/PersonOutlined';
 
 // Internal Dependencies
+import MobileNavItem from './MobileNavItem';
 import presets, { colors } from '../../utils/presets';
-import { rhythm, scale, options } from '../../utils/typography';
-
-// Local Variables
-const texasFlagBlue = '#002868';
+import { options } from '../../utils/typography';
 
 // Component Definition
-const MobileNavItem = ({ linkTo, label, icon: Icon }) => (
-  <Link
-    to={linkTo}
-    css={{
-      color: colors.gray.dark,
-      fontSize: scale(-1 / 2).fontSize,
-      letterSpacing: '0.0075rem',
-      lineHeight: 1,
-      padding: `${rhythm(options.blockMarginBottom / 4)} ${rhythm(
-        options.blockMarginBottom,
-      )} ${rhythm(options.blockMarginBottom / 2)} `,
-      textDecoration: 'none',
-      textAlign: 'center',
-    }}
-  >
-    <div
-      css={{
-        display: 'block',
-        margin: '0 auto',
-      }}
-    >
-      <Icon fill={texasFlagBlue} height="32px" width="32px" />
-    </div>
-    <div
-      css={{
-        marginTop: 8,
-      }}
-    >
-      {label}
-    </div>
-  </Link>
-);
-MobileNavItem.propTypes = {
-  linkTo: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  icon: PropTypes.func.isRequired,
-};
-
 const MobileNav = () => (
   <div
     css={{
@@ -77,9 +31,9 @@ const MobileNav = () => (
       },
     }}
   >
-    <MobileNavItem linkTo="/about/" label="About" icon={InfoOutline} />
+    <MobileNavItem linkTo="/about/" label="About" icon={InfoOutlined} />
     <MobileNavItem linkTo="/resources/" label="Resources" icon={FolderOpen} />
-    <MobileNavItem linkTo="/members/" label="Members" icon={PersonOutline} />
+    <MobileNavItem linkTo="/members/" label="Members" icon={PersonOutlined} />
     <MobileNavItem linkTo="/sponsors/" label="Sponsors" icon={CardGiftCard} />
   </div>
 );
